fix(ops): relocate Stage-2 shadow formulas outside MASTER_V2 surface
Move shadow-only formula/metric producers from src/trading/master_v2 into
the ops shadow-campaign package so the economic boundary Lint Gate remains
fail-closed without technical-wiring exceptions.
