+++ src/ops/productive_pure_stack_numeric_policy_shadow_campaign_v1/fps_atr_or_range_wilder_atr_finalized_ohlcv_v1.py
